har konverterat från int till double så borde funka med decimaltal nu

+++ U210916/U210916X2/Form1.cs
@@ -12,29 +12,29 @@ namespace U210916X2
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int i = int.Parse(textBox1.Text);
-            int j = int.Parse(textBox2.Text);
+            double i = double.Parse(textBox1.Text);
+            double j = double.Parse(textBox2.Text);
             listBox1.Items.Add($"{textBox1.Text}+{textBox2.Text} = {i + j}");
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int i = int.Parse(textBox1.Text);
-            int j = int.Parse(textBox2.Text);
+            double i = double.Parse(textBox1.Text);
+            double j = double.Parse(textBox2.Text);
             listBox1.Items.Add($"{textBox1.Text}+{textBox2.Text} = {i - j}");
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int i = int.Parse(textBox1.Text);
-            int j = int.Parse(textBox2.Text);
+            double i = double.Parse(textBox1.Text);
+            double j = double.Parse(textBox2.Text);
             listBox1.Items.Add($"{textBox1.Text}+{textBox2.Text} = {i * j}");
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int i = int.Parse(textBox1.Text);
-            int j = int.Parse(textBox2.Text);
+            double i = double.Parse(textBox1.Text);
+            double j = double.Parse(textBox2.Text);
             listBox1.Items.Add($"{textBox1.Text}+{textBox2.Text} = {i / j}");
         }
     }
